完善脚本

+++ bj_swarm_deploy/bj_swarm_deploy.sh
@@ -18,20 +18,19 @@ cargo build --release --all
 nohup $violas_path/diem-swarm -c $violascfg_path --diem-node $violas_path/diem-node -n 1 >$violas_scripts_path/swarm.log 2>&1 &
 
 sleep 3
-cd $violascfg_path
-rm -rf $violascfg_path/logs
-cp $violas_scripts_path/start.sh .
-cp $violas_scripts_path/stop.sh .
-cp $violas_scripts_path/cli.sh .
-cp $violas_scripts_path/clean_db_start.sh .
-cp $violas_scripts_path/violas_chain_monitor.py .
-cp $violas_path/diem-node .
-mv mint.key mint_beijing.key
-
-sed -i "89s|level:.*|level: ERROR|g" $violascfg_path/0/node.yaml
-sed -i "108s|address:.*|address: \"0.0.0.0:50001\"|g" $violascfg_path/0/node.yaml
-
-sh stop.sh
-sleep 2
-sh start.sh
-
+if [ ! -d "$violascfg_path" ]; then
+	cd $violascfg_path
+	rm -rf $violascfg_path/logs
+	cp $violas_scripts_path/start.sh .
+	cp $violas_scripts_path/stop.sh .
+	cp $violas_scripts_path/cli.sh .
+	cp $violas_scripts_path/clean_db_start.sh .
+	cp $violas_scripts_path/violas_chain_monitor.py .
+	cp $violas_path/diem-node .
+	mv mint.key mint_beijing.key
+	sed -i "89s|level:.*|level: ERROR|g" $violascfg_path/0/node.yaml
+	sed -i "108s|address:.*|address: \"0.0.0.0:50001\"|g" $violascfg_path/0/node.yaml
+	sh stop.sh
+	sleep 2
+	sh start.sh
+fi
